+++ src/gui/plugins/component_inspector/ComponentInspector.cc
@@ -18,6 +18,7 @@
 #include <iostream>
 #include <list>
 #include <regex>
+#include <unordered_map>
 #include <QColorDialog>
 #include <gz/common/Console.hh>
 #include <gz/common/Profiler.hh>
@@ -72,6 +73,7 @@
 #include "gz/sim/components/Volume.hh"
 #include "gz/sim/components/WindMode.hh"
 #include "gz/sim/components/World.hh"
+#include "gz/sim/config.hh"
 #include "gz/sim/EntityComponentManager.hh"
 #include "gz/sim/gui/GuiEvents.hh"
 
@@ -126,7 +128,33 @@ namespace gz::sim
 
     /// \brief Handles all system info components.
     public: std::unique_ptr<inspector::SystemPluginInfo> systemInfo;
+
+    /// \brief A list of system plugin human readable names.
+    public: QStringList systemNameList;
+
+    /// \brief Maps plugin display names to their filenames.
+    public: std::unordered_map<std::string, std::string> systemMap;
   };
+}
+
+// Helper to remove a prefix from a string if present
+void removePrefix(const std::string &_prefix, std::string &_s)
+{
+  auto id = _s.find(_prefix);
+  if (id != std::string::npos)
+  {
+    _s = _s.substr(_prefix.length());
+  }
+}
+
+// Helper to remove a suffix from a string if present
+void removeSuffix(const std::string &_suffix, std::string &_s)
+{
+  auto id = _s.find(_suffix);
+  if (id != std::string::npos && id + _suffix.length() == _s.length())
+  {
+    _s.erase(id, _suffix.length());
+  }
 }
 
 using namespace gz;
@@ -295,22 +323,22 @@ void gz::sim::setData(QStandardItem *_item,
   _item->setData(QString("Material"),
       ComponentsModel::RoleNames().key("dataType"));
   _item->setData(QList({
-    QVariant(_data.Ambient().R() * 255),
-    QVariant(_data.Ambient().G() * 255),
-    QVariant(_data.Ambient().B() * 255),
-    QVariant(_data.Ambient().A() * 255),
-    QVariant(_data.Diffuse().R() * 255),
-    QVariant(_data.Diffuse().G() * 255),
-    QVariant(_data.Diffuse().B() * 255),
-    QVariant(_data.Diffuse().A() * 255),
-    QVariant(_data.Specular().R() * 255),
-    QVariant(_data.Specular().G() * 255),
-    QVariant(_data.Specular().B() * 255),
-    QVariant(_data.Specular().A() * 255),
-    QVariant(_data.Emissive().R() * 255),
-    QVariant(_data.Emissive().G() * 255),
-    QVariant(_data.Emissive().B() * 255),
-    QVariant(_data.Emissive().A() * 255)
+    QVariant(_data.Ambient().R()),
+    QVariant(_data.Ambient().G()),
+    QVariant(_data.Ambient().B()),
+    QVariant(_data.Ambient().A()),
+    QVariant(_data.Diffuse().R()),
+    QVariant(_data.Diffuse().G()),
+    QVariant(_data.Diffuse().B()),
+    QVariant(_data.Diffuse().A()),
+    QVariant(_data.Specular().R()),
+    QVariant(_data.Specular().G()),
+    QVariant(_data.Specular().B()),
+    QVariant(_data.Specular().A()),
+    QVariant(_data.Emissive().R()),
+    QVariant(_data.Emissive().G()),
+    QVariant(_data.Emissive().B()),
+    QVariant(_data.Emissive().A())
   }), ComponentsModel::RoleNames().key("data"));
 
   // TODO(anyone) Only shows colors of material,
@@ -874,7 +902,6 @@ void ComponentInspector::Update(const UpdateInfo &,
       auto comp = _ecm.Component<components::Material>(this->dataPtr->entity);
       if (comp)
       {
-        this->SetType("material");
         setData(item, comp->Data());
       }
     }
@@ -1091,57 +1118,10 @@ void ComponentInspector::OnMaterialColor(
   double _rDiffuse, double _gDiffuse, double _bDiffuse, double _aDiffuse,
   double _rSpecular, double _gSpecular, double _bSpecular, double _aSpecular,
   double _rEmissive, double _gEmissive, double _bEmissive, double _aEmissive,
-  QString _type, QColor _currColor)
+  QString /*_type*/, QColor /*_currColor*/)
 {
-  // when type is not empty, open qt color dialog
-  std::string type = _type.toStdString();
-  if (!type.empty())
-  {
-    QColor newColor = QColorDialog::getColor(
-        _currColor, nullptr, "Pick a color",
-        {QColorDialog::DontUseNativeDialog, QColorDialog::ShowAlphaChannel});
-
-    // returns if the user hits cancel
-    if (!newColor.isValid())
-      return;
-
-    if (type == "ambient")
-    {
-      _rAmbient = newColor.red();
-      _gAmbient = newColor.green();
-      _bAmbient = newColor.blue();
-      _aAmbient = newColor.alpha();
-    }
-    else if (type == "diffuse")
-    {
-      _rDiffuse = newColor.red();
-      _gDiffuse = newColor.green();
-      _bDiffuse = newColor.blue();
-      _aDiffuse = newColor.alpha();
-    }
-    else if (type == "specular")
-    {
-      _rSpecular = newColor.red();
-      _gSpecular = newColor.green();
-      _bSpecular = newColor.blue();
-      _aSpecular = newColor.alpha();
-    }
-    else if (type == "emissive")
-    {
-      _rEmissive = newColor.red();
-      _gEmissive = newColor.green();
-      _bEmissive = newColor.blue();
-      _aEmissive = newColor.alpha();
-    }
-    else
-    {
-      gzerr << "Invalid material type: " << type << std::endl;
-      return;
-    }
-  }
-
-  std::function<void(const gz::msgs::Boolean &, const bool)> cb =
-      [](const gz::msgs::Boolean &/*_rep*/, const bool _result)
+  std::function<void(const msgs::Boolean &, const bool)> cb =
+      [](const msgs::Boolean &/*_rep*/, const bool _result)
   {
     if (!_result)
       gzerr << "Error setting material color configuration"
@@ -1152,17 +1132,13 @@ void ComponentInspector::OnMaterialColor(
   req.set_id(this->dataPtr->entity);
 
   msgs::Set(req.mutable_material()->mutable_ambient(),
-    math::Color(_rAmbient / 255.0, _gAmbient / 255.0,
-      _bAmbient / 255.0, _aAmbient / 255.0));
+    math::Color(_rAmbient, _gAmbient, _bAmbient, _aAmbient));
   msgs::Set(req.mutable_material()->mutable_diffuse(),
-    math::Color(_rDiffuse / 255.0, _gDiffuse / 255.0,
-      _bDiffuse / 255.0, _aDiffuse / 255.0));
+    math::Color(_rDiffuse, _gDiffuse, _bDiffuse, _aDiffuse));
   msgs::Set(req.mutable_material()->mutable_specular(),
-    math::Color(_rSpecular / 255.0, _gSpecular / 255.0,
-      _bSpecular / 255.0, _aSpecular / 255.0));
+    math::Color(_rSpecular, _gSpecular, _bSpecular, _aSpecular));
   msgs::Set(req.mutable_material()->mutable_emissive(),
-    math::Color(_rEmissive / 255.0, _gEmissive / 255.0,
-      _bEmissive / 255.0, _aEmissive / 255.0));
+    math::Color(_rEmissive, _gEmissive, _bEmissive, _aEmissive));
 
   auto materialCmdService = "/world/" + this->dataPtr->worldName
       + "/visual_config";
@@ -1229,6 +1205,106 @@ const std::string &ComponentInspector::WorldName() const
 transport::Node &ComponentInspector::TransportNode()
 {
   return this->dataPtr->node;
+}
+
+/////////////////////////////////////////////////
+void ComponentInspector::QuerySystems()
+{
+  msgs::Empty req;
+  msgs::EntityPlugin_V res;
+  bool result;
+  unsigned int timeout = 5000;
+  std::string service{"/world/" + this->dataPtr->worldName +
+      "/system/info"};
+  if (!this->dataPtr->node.Request(service, req, timeout, res, result))
+  {
+    ignerr << "Unable to query available systems." << std::endl;
+    return;
+  }
+
+  this->dataPtr->systemNameList.clear();
+  this->dataPtr->systemMap.clear();
+  for (const auto &plugin : res.plugins())
+  {
+    if (plugin.filename().empty())
+    {
+      ignerr << "Received empty plugin name. This shouldn't happen."
+             << std::endl;
+      continue;
+    }
+
+    // Remove common prefixes and suffixes
+    auto humanReadable = plugin.filename();
+    removePrefix("ignition-gazebo-", humanReadable);
+    removePrefix("ignition-gazebo" +
+        std::string(GZ_SIM_MAJOR_VERSION_STR) + "-", humanReadable);
+    removeSuffix("-system", humanReadable);
+    removeSuffix("system", humanReadable);
+    removeSuffix("-plugin", humanReadable);
+    removeSuffix("plugin", humanReadable);
+
+    // Replace - with space, capitalize
+    std::replace(humanReadable.begin(), humanReadable.end(), '-', ' ');
+    humanReadable[0] = std::toupper(humanReadable[0]);
+
+    this->dataPtr->systemMap[humanReadable] = plugin.filename();
+    this->dataPtr->systemNameList.push_back(
+        QString::fromStdString(humanReadable));
+  }
+  this->dataPtr->systemNameList.sort();
+  this->dataPtr->systemNameList.removeDuplicates();
+  this->SystemNameListChanged();
+}
+
+/////////////////////////////////////////////////
+QStringList ComponentInspector::SystemNameList() const
+{
+  return this->dataPtr->systemNameList;
+}
+
+/////////////////////////////////////////////////
+void ComponentInspector::SetSystemNameList(const QStringList &_list)
+{
+  this->dataPtr->systemNameList = _list;
+}
+
+/////////////////////////////////////////////////
+void ComponentInspector::OnAddSystem(const QString &_name,
+    const QString &_filename, const QString &_innerxml)
+{
+  auto filenameStr = _filename.toStdString();
+  auto it = this->dataPtr->systemMap.find(filenameStr);
+  if (it == this->dataPtr->systemMap.end())
+  {
+    ignerr << "Internal error: failed to find [" << filenameStr
+           << "] in system map." << std::endl;
+    return;
+  }
+
+  msgs::EntityPlugin_V req;
+  auto ent = req.mutable_entity();
+  ent->set_id(this->dataPtr->entity);
+  auto plugin = req.add_plugins();
+  std::string name = _name.toStdString();
+  std::string filename = this->dataPtr->systemMap[filenameStr];
+  std::string innerxml = _innerxml.toStdString();
+  plugin->set_name(name);
+  plugin->set_filename(filename);
+  plugin->set_innerxml(innerxml);
+
+  msgs::Boolean res;
+  bool result;
+  unsigned int timeout = 5000;
+  std::string service{"/world/" + this->dataPtr->worldName +
+      "/entity/system/add"};
+  if (!this->dataPtr->node.Request(service, req, timeout, res, result))
+  {
+    ignerr << "Error adding new system to entity: "
+           << this->dataPtr->entity << "\n"
+           << "Name: " << name << "\n"
+           << "Filename: " << filename << "\n"
+           << "Inner XML: " << innerxml << std::endl;
+  }
 }
 
 // Register this plugin